EscritoAAudio/requirements.txt
paddlepaddle
paddleocr
torch
transformers
scipy
soundfile
opencv-python
pillow
